+++ packages/analytics/analytics-chart/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@0.13.2...@kong-ui-public/analytics-chart@0.13.3) (2023-10-27)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
+## [0.13.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@0.13.1...@kong-ui-public/analytics-chart@0.13.2) (2023-10-25)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
+## [0.13.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@0.13.0...@kong-ui-public/analytics-chart@0.13.1) (2023-10-24)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
+# [0.13.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@0.12.3...@kong-ui-public/analytics-chart@0.13.0) (2023-10-24)
+
+
+### Features
+
+* **sandbox-layout:** create shared component sandbox layout ([#860](https://github.com/Kong/public-ui-components/issues/860)) ([0bb9550](https://github.com/Kong/public-ui-components/commit/0bb95505c4c86e172b84514c6a984e1d158be397))
+
+
+
+
+
 ## [0.12.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@0.12.2...@kong-ui-public/analytics-chart@0.12.3) (2023-10-24)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-chart