+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.14](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@1.4.13...@kong-ui-public/entities-plugins@1.4.14) (2023-10-27)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [1.4.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@1.4.12...@kong-ui-public/entities-plugins@1.4.13) (2023-10-25)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [1.4.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@1.4.11...@kong-ui-public/entities-plugins@1.4.12) (2023-10-25)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [1.4.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@1.4.10...@kong-ui-public/entities-plugins@1.4.11) (2023-10-24)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [1.4.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@1.4.9...@kong-ui-public/entities-plugins@1.4.10) (2023-10-24)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
 ## [1.4.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@1.4.8...@kong-ui-public/entities-plugins@1.4.9) (2023-10-24)
 
 **Note:** Version bump only for package @kong-ui-public/entities-plugins