+++ packages/entities/entities-snis/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.43](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@1.1.42...@kong-ui-public/entities-snis@1.1.43) (2023-10-27)
+
+**Note:** Version bump only for package @kong-ui-public/entities-snis
+
+
+
+
+
+## [1.1.42](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@1.1.41...@kong-ui-public/entities-snis@1.1.42) (2023-10-25)
+
+**Note:** Version bump only for package @kong-ui-public/entities-snis
+
+
+
+
+
+## [1.1.41](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@1.1.40...@kong-ui-public/entities-snis@1.1.41) (2023-10-24)
+
+**Note:** Version bump only for package @kong-ui-public/entities-snis
+
+
+
+
+
+## [1.1.40](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@1.1.39...@kong-ui-public/entities-snis@1.1.40) (2023-10-24)
+
+**Note:** Version bump only for package @kong-ui-public/entities-snis
+
+
+
+
+
 ## [1.1.39](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@1.1.38...@kong-ui-public/entities-snis@1.1.39) (2023-10-24)
 
 **Note:** Version bump only for package @kong-ui-public/entities-snis